float to int

+++ src/main/java/logic/CreateurAstre.java
@@ -1,5 +1,5 @@
 package logic;
 
 public abstract  class CreateurAstre {
-    public abstract Astre factory(String nom, float taille, float masse, int pX, int pY);
+    public abstract Astre factory(String nom, int taille, int masse, int pX, int pY);
 }
