+++ src/main/java/logic/Planete.java
@@ -1,14 +1,14 @@
 package logic;
 
 public class Planete extends Astre{
-    private float taille;
-    private float masse;
+    private int taille;
+    private int masse;
     private String nom;
     private int positionX;
     private int positionY;
 
 
-    public Planete(String nom, float taille, float masse, int pX, int pY ) {
+    public Planete(String nom, int taille, int masse, int pX, int pY ) {
         this.taille = taille;
         this.masse = masse;
         this.nom = nom;
